Add a nix-based build environment
While experimenting with the repository, I've found that it's hard to
build. The list of dependencies is implicit and it takes time to
actually get the stuff compiled and run properly.

I'm not sure if anyone else would use this, but since we took time
to freeze the dependencies, I'm still submitting the result as a
pull-request, at least so that it can be used as a reference.